+++ rust/src/router/data_consumer.rs
@@ -15,7 +15,7 @@ use crate::worker::{Channel, NotificationParseError, RequestError, SubscriptionH
 use async_executor::Executor;
 use event_listener_primitives::{Bag, BagOnce, HandlerId};
 use log::{debug, error};
-use mediasoup_sys::fbs::{data_consumer, data_producer, notification, response};
+use mediasoup_sys::fbs::{data_consumer, notification, response};
 use mediasoup_types::data_structures::{AppData, WebRtcMessage};
 use mediasoup_types::sctp_parameters::SctpStreamParameters;
 use parking_lot::Mutex;
@@ -157,6 +157,7 @@ pub struct DataConsumerDump {
     pub label: String,
     pub protocol: String,
     pub sctp_stream_parameters: Option<SctpStreamParameters>,
+    pub buffered_amount: u32,
     pub buffered_amount_low_threshold: u32,
     pub paused: bool,
     pub subchannels: Vec<u16>,
@@ -171,7 +172,7 @@ impl<'a> TryFromFbs<'a> for DataConsumerDump {
         Ok(Self {
             id: dump.id.parse()?,
             data_producer_id: dump.data_producer_id.parse()?,
-            r#type: if dump.type_ == data_producer::Type::Sctp {
+            r#type: if dump.type_ == data_consumer::Type::Sctp {
                 DataConsumerType::Sctp
             } else {
                 DataConsumerType::Direct
@@ -182,6 +183,7 @@ impl<'a> TryFromFbs<'a> for DataConsumerDump {
                 .sctp_stream_parameters
                 .as_ref()
                 .map(|parameters| SctpStreamParameters::from_fbs(parameters.as_ref())),
+            buffered_amount: dump.buffered_amount,
             buffered_amount_low_threshold: dump.buffered_amount_low_threshold,
             paused: dump.paused,
             subchannels: dump.subchannels.clone(),