+++ worker/include/RTC/Router.hpp
@@ -2,7 +2,6 @@
 #define MS_RTC_ROUTER_HPP
 
 #include "common.hpp"
-#include "SharedInterface.hpp"
 #include "Channel/ChannelRequest.hpp"
 #include "RTC/Consumer.hpp"
 #include "RTC/DataConsumer.hpp"
@@ -14,6 +13,7 @@
 #include "RTC/SCTP/public/Message.hpp"
 #include "RTC/Transport.hpp"
 #include "RTC/WebRtcServer.hpp"
+#include "SharedInterface.hpp"
 #include <absl/container/flat_hash_map.h>
 #include <absl/container/flat_hash_set.h>
 #include <string>
@@ -49,8 +49,8 @@ namespace RTC
 		void HandleRequest(Channel::ChannelRequest* request) override;
 
 	private:
-		RTC::Transport* GetTransportById(const std::string& transportId) const;
-		RTC::RtpObserver* GetRtpObserverById(const std::string& rtpObserverId) const;
+		RTC::Transport* AssertAndGetTransportById(const std::string& transportId) const;
+		RTC::RtpObserver* AssertAndGetRtpObserverById(const std::string& rtpObserverId) const;
 		void CheckNoTransport(const std::string& transportId) const;
 		void CheckNoRtpObserver(const std::string& rtpObserverId) const;
 
@@ -94,15 +94,6 @@ namespace RTC
 		void OnTransportDataProducerPaused(RTC::Transport* transport, RTC::DataProducer* dataProducer) override;
 		void OnTransportDataProducerResumed(
 		  RTC::Transport* transport, RTC::DataProducer* dataProducer) override;
-		// TODO: SCTP: Remove when we migrate to the new SCTP stack.
-		void OnTransportDataProducerMessageReceived(
-		  RTC::Transport* transport,
-		  RTC::DataProducer* dataProducer,
-		  const uint8_t* msg,
-		  size_t len,
-		  uint32_t ppid,
-		  std::vector<uint16_t>& subchannels,
-		  std::optional<uint16_t> requiredSubchannel) override;
 		void OnTransportDataProducerMessageReceived(
 		  RTC::Transport* transport,
 		  RTC::DataProducer* dataProducer,