+++ worker/include/Settings.hpp
@@ -2,8 +2,8 @@
 #define MS_SETTINGS_HPP
 
 #include "common.hpp"
-#include "LogLevel.hpp"
 #include "Channel/ChannelRequest.hpp"
+#include "LogLevel.hpp"
 #include <absl/container/flat_hash_map.h>
 #include <string>
 #include <vector>
@@ -40,7 +40,6 @@ public:
 		std::string dtlsPrivateKeyFile;
 		std::string libwebrtcFieldTrials{ "WebRTC-Bwe-AlrLimitedBackoff/Enabled/" };
 		bool disableLiburing{ false };
-		bool useBuiltInSctpStack{ false };
 	};
 
 public: